requirements.txt
pandas
httpx[http2]
python-dotenv